DimeAPI/requirements/requirements.txt
argh==0.26.2
asn1crypto==0.24.0
apscheduler==3.5.1
Babel==2.5.3
base58==0.2.5
bcrypt==3.1.3
bleach==1.5.0
certifi==2018.1.18
cffi==1.11.2
chardet==3.0.4
click==6.7
cryptography==2.1.4
cytoolz==0.9.0
Django==2.0.1
django-cors-headers==2.1.0
django-cors-middleware==1.3.1
django-debug-toolbar==1.9
django-filter==1.0.4
django-ipfs-storage==0.0.4
django-ipware==1.1.6
django-oauth-toolkit==1.0.0
django-reset-migrations==0.3.1
django-sslserver==0.20
django-two-factor-auth==1.7.0
django-unixdatetimefield==0.1.6
django-user-agents==0.3.2
djangorestframework==3.7.7
djangorestframework-jsonapi==2.3.1
djangorestframework-jwt==1.11.0
djangorestframework-oauth==1.1.0
enum34==1.1.6
eth-abi==0.5.0
eth-keyfile==0.4.1
eth-keys==0.1.0b4
eth-tester==0.1.0b12
eth-utils==0.7.4
ethereum==2.3.0
ExifRead==2.1.2
Flask==0.12.2
future==0.16.0
gevent==1.2.2
gipc==0.6.0
googlemaps==2.5.1
greenlet==0.4.12
html5lib==0.9999999
idna==2.6
inflection==0.3.1
inifile==0.4
ipfs-api==0.2.3
ipfsapi==0.4.2.post1
itsdangerous==0.24
Jinja2==2.10
Lektor==3.0.1
lru-dict==1.1.6
Markdown==2.6.10
MarkupSafe==1.0
mistune==0.8.3
multihash==0.1.1
mysql-connector-python==8.0.5
mysqlclient==1.3.12
numpy==1.13.3
oauthlib==2.0.6
olefile==0.44
django-otp==0.4.2
pathtools==0.1.2
Pillow==4.3.0
protobuf==3.5.0.post1
py-ecc==1.4.2
pycparser==2.18
pycryptodome==3.4.7
PyJWT==1.5.3
pylru==1.0.9
PyMySQL==0.8.0
pyOpenSSL==17.5.0
pysha3==1.0.2
pytz==2017.3
PyYAML==3.12
repoze.lru==0.7
requests==2.18.4
rlp==0.6.0
semantic-version==2.6.0
six==1.11.0
sqlparse==0.2.4
style==1.1.0
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0rc3
toolz==0.9.0
ua-parser==0.7.3
update==0.0.1
urllib3==1.22
user-agents==1.1.0
watchdog==0.8.3
web3==4.0.0b6
Werkzeug==0.13
tzlocal==1.5.1